add __CODE__RED__GLOBAL__INCLUDE__

+++ CodeRed/DirectX12/DirectX12DisplayAdapter.cpp
@@ -5,8 +5,8 @@
 CodeRed::DirectX12DisplayAdapter::DirectX12DisplayAdapter(
 	const WRL::ComPtr<IDXGIAdapter1> &adapter,
 	const std::string &name,
-	const UInt32 device_id,
-	const UInt32 vendor_id) :
+	const size_t device_id,
+	const size_t vendor_id) :
 	GpuDisplayAdapter(name, device_id, vendor_id),
 	mAdapter(adapter)
 {
