+++ CodeRed/Shared/Information/WindowInfo.hpp
@@ -1,12 +1,10 @@
 #pragma once
 
-#include <Shared/Utility.hpp>
-
 namespace CodeRed {
 
 	struct WindowInfo {
-		UInt32 width;
-		UInt32 height;
+		size_t width;
+		size_t height;
 
 		void* handle;
 	};
